Results of progress in semantic analysis written in main.rs, still unfinished.

+++ src/Semantic/semantic_rules.rs
@@ -33,7 +33,7 @@ impl SemanticRules {
     pub fn validate_array_declaration(
         name: &str,
         element_type: &Types, 
-        size: usize
+        size: i16
     ) -> Result<(), String> {
         // Check array name length
         if name.len() > 8 {
